Reorganized firebase
