+++ app/firebase/firebase.js
@@ -1,0 +1,24 @@
+import firebase from 'firebase'
+require('firebase/firestore')
+
+// Initialize Firebase
+const config = {
+  apiKey: process.env.FIREBASE_API_KEY,
+  authDomain: process.env.FIREBASE_AUTH_DOMAIN,
+  databaseURL: process.env.FIREBASE_DATABASE_URL,
+  projectId: process.env.FIREBASE_PROJECT_ID,
+  storageBucket: process.env.FIREBASE_STORAGE_BUCKET,
+  messagingSenderId: process.env.FIREBASE_MESSAGING_SENDER_ID
+}
+
+if (!firebase.apps.length) {
+  firebase.initializeApp(config)
+}
+
+const auth = firebase.auth()
+const googleAuthProvider = new firebase.auth.GoogleAuthProvider()
+
+export {
+  auth,
+  googleAuthProvider
+}
