+++ app/firebase/index.js
@@ -1,22 +1,7 @@
-import firebase from 'firebase'
-require('firebase/firestore')
+import * as auth from './auth'
+import * as firebase from './firebase'
 
-// Initialize Firebase
-const config = {
-  apiKey: process.env.FIREBASE_API_KEY,
-  authDomain: process.env.FIREBASE_AUTH_DOMAIN,
-  databaseURL: process.env.FIREBASE_DATABASE_URL,
-  projectId: process.env.FIREBASE_PROJECT_ID,
-  storageBucket: process.env.FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: process.env.FIREBASE_MESSAGING_SENDER_ID
+export {
+  auth,
+  firebase
 }
-
-if (!firebase.apps.length) {
-  firebase.initializeApp(config)
-}
-
-export const db = firebase.firestore()
-
-// at the bottom of src/firebase.js
-export const auth = firebase.auth()
-export const googleAuthProvider = new firebase.auth.GoogleAuthProvider()
